fix: patch update, generate CldImageProps.svelte.d.ts
