+++ svelte-cloudinary/package.json
@@ -53,10 +53,10 @@
 		"svelte-preprocess": "^5.0.3",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
+		"unpic": "3.6.1",
 		"vite": "^4.3.0",
 		"vitest": "^0.25.3"
 	},
-	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
 	"type": "module",
 	"dependencies": {
