+++ svelte-cloudinary/src/lib/components/CldOgImage.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { getCldImageUrl } from '$lib/helpers/getCldImageUrl.js';
 	import type { ImageOptions } from '@cloudinary-util/url-loader';
-	import type { CldImageProps } from './CldImage.svelte';
+	import type { CldImageProps } from './CldImageTypes.ts';
 	import { OG_IMAGE_WIDTH, OG_IMAGE_WIDTH_RESIZE, OG_IMAGE_HEIGHT } from '$lib/constants/sizes.js';
 
 	const TWITTER_CARD = 'summary_large_image';
