--- svelte-cloudinary/src/lib/index.js
@@ -1,12 +1,0 @@
-// Reexport your entry components here
-import CldImage from './components/CldImage.svelte'
-export { CldImage }
-
-import CldOgImage from './components/CldOgImage.svelte'
-export { CldOgImage }
-
-import CldUploadWidget from './components/CldUploadWidget.svelte'
-export { CldUploadWidget} 
-
-import CldUploadButton from './components/CldUploadButton.svelte'
-export { CldUploadButton}
