+++ svelte-cloudinary/src/lib/index.ts
@@ -1,3 +1,17 @@
+// Reexport your entry components here
+import CldImage from './components/CldImage.svelte'
+export { CldImage }
+export type { CldImageProps } from "./types.ts";
+
+import CldOgImage from './components/CldOgImage.svelte'
+export { CldOgImage }
+
+import CldUploadWidget from './components/CldUploadWidget.svelte'
+export { CldUploadWidget}
+
+import CldUploadButton from './components/CldUploadButton.svelte'
+export { CldUploadButton}
+
 export { getCldImageUrl } from '$lib/helpers/getCldImageUrl.ts';
 export type { GetCldImageUrl, GetCldImageUrlOptions, GetCldImageUrlConfig, GetCldImageUrlAnalytics } from '$lib/helpers/getCldImageUrl.ts';
 
@@ -6,4 +20,3 @@ export type { GetCldOgImageUrl } from '$lib/helpers/getCldOgImageUrl.ts';
 
 export type { CloudinaryVideoPlayer, CloudinaryVideoPlayerOptions, CloudinaryVideoPlayerOptionsColors, CloudinaryVideoPlayerOptionsLogo } from './types/player.ts';
 
-export { default as CldImage } from '$lib/components/CldImage.svelte';
